에픽 목록 조회 API 구현

See merge request s12-final/S12P31D207!49

+++ backend/core/core-api/src/main/java/com/ssafy/taskit/api/controller/EpicController.java
@@ -42,8 +42,8 @@ public class EpicController {
   public ApiResponse<List<EpicResponse>> findEpics(ApiUser apiUser, @PathVariable Long projectId) {
     List<Epic> epics = epicService.findEpics(apiUser.toUser(), projectId);
     List<Long> epicIds = epics.stream().map(Epic::id).toList();
-    Map<Long, Integer> totalIssueCountMap = issueService.countTotalIssues(epicIds);
-    Map<Long, Integer> backlogIssueCountMap = issueService.countBacklogIssues(epicIds);
+    Map<Long, Long> totalIssueCountMap = issueService.countTotalIssues(epicIds);
+    Map<Long, Long> backlogIssueCountMap = issueService.countBacklogIssues(epicIds);
     List<EpicResponse> response =
         EpicResponse.from(epics, totalIssueCountMap, backlogIssueCountMap);
     return ApiResponse.success(response);
